+++ apps/docs/src/components/Inputs/switch/SizeView.tsx
@@ -1,4 +1,5 @@
-import { Switch, VBox } from "@elf-framework/ui";
+import React from "react";
+import { Switch, VBox } from "@react-elf/ui";
 
 export function SizeView() {
   return (
@@ -8,6 +9,9 @@ export function SizeView() {
         backgroundColor: "var(--color-gray-0)",
         padding: [60, 80],
         fontSize: 13,
+        justifyContent: "center",
+        alignItems: "center",
+        height: 300,
       }}
     >
       <div
@@ -37,7 +41,7 @@ export function SizeView() {
         </div>
         <div style={{ display: "flex", gap: 10 }}>
           <em style={{ display: "inline-block", width: 100 }}>Extra large</em>
-          <Switch size="extra-large" checked>
+          <Switch size="extra-large" defaultChecked>
             Label
           </Switch>
         </div>