+++ packages/@react-elf/shared/src/Icon.tsx
@@ -1,5 +1,5 @@
-import { CommonStyle, KeyValue } from "@react-elf-types/shared";
 import React from "react";
+import { CommonStyle, KeyValue } from "@react-elf-types/shared";
 
 interface IconWrapperProps {
   children: React.ReactNode;