+++ packages/@react-elf/ui/src/index.ts
@@ -20,3 +20,4 @@ export * from "@react-elf/checkbox";
 export * from "@react-elf/radio";
 export * from "@react-elf/help-text";
 export * from "@react-elf/input-editor";
+export * from "@react-elf/switch";