+++ packages/@react-elf/ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @react-elf/button
 
+## 0.0.88
+
+### Patch Changes
+
+- add Switch component
+
+- Updated dependencies []:
+  - @react-elf/input-editor@0.0.88
+  - @react-elf/shared@0.0.88
+  - @react-elf/switch@0.0.88
+  - @react-elf/alert@0.0.88
+  - @react-elf/badge@0.0.88
+  - @react-elf/button@0.0.88
+  - @react-elf/checkbox@0.0.88
+  - @react-elf/color-mixer@0.0.88
+  - @react-elf/dialog@0.0.88
+  - @react-elf/divider@0.0.88
+  - @react-elf/ghost@0.0.88
+  - @react-elf/help-text@0.0.88
+  - @react-elf/layout@0.0.88
+  - @react-elf/menu@0.0.88
+  - @react-elf/notification@0.0.88
+  - @react-elf/panel@0.0.88
+  - @react-elf/popover@0.0.88
+  - @react-elf/radio@0.0.88
+  - @react-elf/toast@0.0.88
+  - @react-elf/toolbar@0.0.88
+  - @react-elf/tools@0.0.88
+  - @react-elf/tooltip@0.0.88
+
 ## 0.0.87
 
 ### Patch Changes